requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
lxml==5.2.1
openai==1.23.2
pillow==10.3.0
pydantic==2.7.0
pydantic_core==2.18.1
python-docx==1.1.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.11.0
